Fix order amount types and update tests

+++ order-service/src/main/java/com/loadup/orderservice/dto/CreateOrderRequest.java
@@ -4,8 +4,10 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record CreateOrderRequest(
         @NotBlank(message = "customerId is required") String customerId,
         @NotNull(message = "totalAmount is required")
-        @Positive(message = "totalAmount must be positive") Double totalAmount
+        @Positive(message = "totalAmount must be positive") BigDecimal totalAmount
 ) {}
